Rename utils.py to avoid naming collision
